osf/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
Django==1.6.11
django-filebrowser==3.5.7
django-grappelli==2.5.7
django-guardian==1.2.5
django-userena==1.4.1
djangorestframework==3.1.2
easy-thumbnails==2.2
html2text==2014.12.29
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==2.8.1
pylibmc==1.4.3
pytz==2015.4
qiniu==7.0.5
simplejson==3.7.2
six==1.9.0
South==1.0.2
